Files were reordered and the dependency injection using hilt was finalized.
