+++ app/src/main/java/gonzalo/dev/marvelapp/common/annotation/Timeout.kt
@@ -1,0 +1,5 @@
+package gonzalo.dev.marvelapp.common.annotation
+
+@Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.CLASS)
+annotation class Timeout(val timeout: Long)
