+++ app/src/main/java/gonzalo/dev/marvelapp/home/data/dto/CharacterResponse.kt
@@ -1,0 +1,18 @@
+package gonzalo.dev.marvelapp.home.data.dto
+
+data class CharacterResponse(
+    val code: Int,
+    val status: String,
+    val copyright: String,
+    val attributionText: String,
+    val attributionHTML: String,
+    val data: Data
+)
+
+data class Data(
+    val offset: Int,
+    val limit: Int,
+    val total: Int,
+    val count: Int,
+    val results: List<CharacterData>
+)
