+++ app/src/main/java/gonzalo/dev/marvelapp/home/data/source/CharacterDataSource.kt
@@ -1,0 +1,12 @@
+package gonzalo.dev.marvelapp.home.data.source
+
+import gonzalo.dev.marvelapp.home.data.dto.CharacterResponse
+import kotlinx.coroutines.flow.Flow
+
+interface CharacterDataSource {
+    fun fetchCharacters(
+        offset: Int?,
+        publicKey: String,
+        privateKey: String
+    ): Flow<CharacterResponse>
+}
