+++ app/src/main/res/layout/activity_base.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/baseContainer"
-    tools:context=".home.HomeActivity">
+    tools:context=".home.presentation.HomeActivity">
 
     <FrameLayout
         android:id="@+id/base_layout_view"
