--- core/src/main/java/gonzalo/dev/core/common/Timeout.kt
@@ -1,5 +1,0 @@
-package gonzalo.dev.core.common
-
-@Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.CLASS)
-annotation class Timeout(val timeout: Long)
