--- core/src/main/java/gonzalo/dev/core/data/dto/CharacterResponse.kt
@@ -1,18 +1,0 @@
-package gonzalo.dev.core.data.dto
-
-data class CharacterResponse(
-    val code: Int,
-    val status: String,
-    val copyright: String,
-    val attributionText: String,
-    val attributionHTML: String,
-    val data: Data
-)
-
-data class Data(
-    val offset: Int,
-    val limit: Int,
-    val total: Int,
-    val count: Int,
-    val results: List<CharacterData>
-)
