--- core/src/main/java/gonzalo/dev/core/data/repository/CharacterRepository.kt
@@ -1,13 +1,0 @@
-package gonzalo.dev.core.data.repository;
-
-import gonzalo.dev.core.data.dto.CharacterResponse
-import kotlinx.coroutines.flow.Flow
-
-interface CharacterRepository {
-
-    fun fetchCharacters(
-        offset: Int?,
-        publicKey: String,
-        privateKey: String
-    ): Flow<CharacterResponse>
-}
