+++ settings.gradle
@@ -1,3 +1,2 @@
 rootProject.name = "MarvelApp"
 include ':app'
-include ':core'
